Add two new APIs to split the upgrade process
PrepareAndCheckBinary() downloads the binary to a location different
to the final destination, it also applies the patch and does the
checksum if configured.

CommitBinary() will move the staging binary to the final location.

This is useful when you want to upgrade many machines at the same time
but you want to load the new binary only when the download is successful
in all of those machines. In that case, CommitBinary() will less likely
to fail.

Apply() will keep its current behavior and will call
PrepareAndCheckBinary() and CommitBinary()